+++ core/api/typescript/src/kurtosis_core_rpc_api_bindings/api_container_service_pb.js
@@ -3184,8 +3184,8 @@ proto.api_container_api.StartServiceArgs.toObject = function(includeInstance, ms
     dockerEnvVarsMap: (f = msg.getDockerEnvVarsMap()) ? f.toObject(includeInstance, undefined) : [],
     filesArtifactMountpointsMap: (f = msg.getFilesArtifactMountpointsMap()) ? f.toObject(includeInstance, undefined) : [],
     publicPortsMap: (f = msg.getPublicPortsMap()) ? f.toObject(includeInstance, proto.api_container_api.Port.toObject) : [],
-    cpuAllocation: jspb.Message.getFieldWithDefault(msg, 10, ""),
-    memoryAllocation: jspb.Message.getFieldWithDefault(msg, 11, 0)
+    cpuAllocationMillicpus: jspb.Message.getFieldWithDefault(msg, 10, 0),
+    memoryAllocationMegabytes: jspb.Message.getFieldWithDefault(msg, 11, 0)
   };
 
   if (includeInstance) {
@@ -3263,12 +3263,12 @@ proto.api_container_api.StartServiceArgs.deserializeBinaryFromReader = function(
          });
       break;
     case 10:
-      var value = /** @type {string} */ (reader.readString());
-      msg.setCpuAllocation(value);
+      var value = /** @type {number} */ (reader.readUint64());
+      msg.setCpuAllocationMillicpus(value);
       break;
     case 11:
       var value = /** @type {number} */ (reader.readUint64());
-      msg.setMemoryAllocation(value);
+      msg.setMemoryAllocationMegabytes(value);
       break;
     default:
       reader.skipField();
@@ -3343,14 +3343,14 @@ proto.api_container_api.StartServiceArgs.serializeBinaryToWriter = function(mess
   if (f && f.getLength() > 0) {
     f.serializeBinary(9, writer, jspb.BinaryWriter.prototype.writeString, jspb.BinaryWriter.prototype.writeMessage, proto.api_container_api.Port.serializeBinaryToWriter);
   }
-  f = message.getCpuAllocation();
-  if (f.length > 0) {
-    writer.writeString(
+  f = message.getCpuAllocationMillicpus();
+  if (f !== 0) {
+    writer.writeUint64(
       10,
       f
     );
   }
-  f = message.getMemoryAllocation();
+  f = message.getMemoryAllocationMegabytes();
   if (f !== 0) {
     writer.writeUint64(
       11,
@@ -3559,28 +3559,28 @@ proto.api_container_api.StartServiceArgs.prototype.clearPublicPortsMap = functio
 
 
 /**
- * optional string cpu_allocation = 10;
- * @return {string}
- */
-proto.api_container_api.StartServiceArgs.prototype.getCpuAllocation = function() {
-  return /** @type {string} */ (jspb.Message.getFieldWithDefault(this, 10, ""));
-};
-
-
-/**
- * @param {string} value
- * @return {!proto.api_container_api.StartServiceArgs} returns this
- */
-proto.api_container_api.StartServiceArgs.prototype.setCpuAllocation = function(value) {
-  return jspb.Message.setProto3StringField(this, 10, value);
-};
-
-
-/**
- * optional uint64 memory_allocation = 11;
+ * optional uint64 cpu_allocation_millicpus = 10;
  * @return {number}
  */
-proto.api_container_api.StartServiceArgs.prototype.getMemoryAllocation = function() {
+proto.api_container_api.StartServiceArgs.prototype.getCpuAllocationMillicpus = function() {
+  return /** @type {number} */ (jspb.Message.getFieldWithDefault(this, 10, 0));
+};
+
+
+/**
+ * @param {number} value
+ * @return {!proto.api_container_api.StartServiceArgs} returns this
+ */
+proto.api_container_api.StartServiceArgs.prototype.setCpuAllocationMillicpus = function(value) {
+  return jspb.Message.setProto3IntField(this, 10, value);
+};
+
+
+/**
+ * optional uint64 memory_allocation_megabytes = 11;
+ * @return {number}
+ */
+proto.api_container_api.StartServiceArgs.prototype.getMemoryAllocationMegabytes = function() {
   return /** @type {number} */ (jspb.Message.getFieldWithDefault(this, 11, 0));
 };
 
@@ -3589,7 +3589,7 @@ proto.api_container_api.StartServiceArgs.prototype.getMemoryAllocation = functio
  * @param {number} value
  * @return {!proto.api_container_api.StartServiceArgs} returns this
  */
-proto.api_container_api.StartServiceArgs.prototype.setMemoryAllocation = function(value) {
+proto.api_container_api.StartServiceArgs.prototype.setMemoryAllocationMegabytes = function(value) {
   return jspb.Message.setProto3IntField(this, 11, value);
 };