+++ core/api/typescript/src/lib/constructor_calls.ts
@@ -36,7 +36,6 @@ import {
 import { ServiceID } from './services/service';
 import { PartitionID } from './enclaves/enclave_context';
 import { ModuleID } from "./modules/module_context";
-import {c} from "tar";
 
 // ==============================================================================================
 //                           Shared Objects (Used By Multiple Endpoints)
@@ -156,8 +155,8 @@ export function newStartServiceArgs(
     cmdArgs: string[],
     dockerEnvVars: Map<string, string>,
     filesArtifactMountDirpaths: Map<string, string>,
-    cpuAllocation: string,
-    memoryAllocation: number,
+    cpuAllocationMillicpus: number,
+    memoryAllocationMegabytes: number,
 ): StartServiceArgs {
     const result: StartServiceArgs = new StartServiceArgs();
     result.setServiceId(String(serviceId));
@@ -182,6 +181,8 @@ export function newStartServiceArgs(
     for (const [artifactId, mountDirpath] of filesArtifactMountDirpaths.entries()) {
         filesArtificatMountDirpathsMap.set(artifactId, mountDirpath);
     }
+    result.setCpuAllocationMillicpus(cpuAllocationMillicpus);
+    result.setMemoryAllocationMegabytes(memoryAllocationMegabytes);
 
     //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
     const publicPortsMap: jspb.Map<string, Port> = result.getPublicPortsMap();
@@ -189,8 +190,7 @@ export function newStartServiceArgs(
         publicPortsMap.set(portId, portSpec);
     }
     //TODO finish the hack
-    result.setCpuAllocation(cpuAllocation);
-    result.setMemoryAllocation(memoryAllocation);
+
     return result;
 }