+++ core/api/typescript/src/lib/enclaves/enclave_context.ts
@@ -292,8 +292,8 @@ export class EnclaveContext {
             containerConfig.cmdOverrideArgs,
             containerConfig.environmentVariableOverrides,
             artifactIdStrToMountDirpath,
-            containerConfig.cpuAllocation,
-            containerConfig.memoryAllocation,
+            containerConfig.cpuAllocationMillicpus,
+            containerConfig.memoryAllocationMegabytes,
         );
 
         const startServiceResponseResult = await this.backend.startService(startServiceArgs)